+++ modules/micropython.cmake
@@ -1,6 +1,7 @@
 # Auto-generated - include all compiled modules
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_algorithms/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_bitwise/micropython.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/usermod_builtins_demo/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_counter/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_dict_operations/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_factorial/micropython.cmake)